refactor(reconcile): extract shared reconcile package (Collection, PersistIssueDetails, Extractor)
Step 2 of the round-17 detail-outcome/swrRefresh design
(docs/plans/2026-07-08-detail-outcome-swr-refresh-design.md). New package
internal/reconcile, shared by internal/sync today and internal/repo in a
later step (verified acyclic — neither imports the other):

- syncCollection moves from internal/sync as reconcile.Collection with the
  same semantics (upsert-all log-and-continue, prune runs once iff every
  upsert returned nil, nil prune = upsert-only), plus one interface
  addition: it returns clean bool (true iff all upserts succeeded; a prune
  failure doesn't affect it). All 7 worker metadata sites switch over and
  ignore the return — a later step consumes it. Moved tests pin the return.

- PersistIssueDetails hosts the five per-issue collection specs (comments,
  documents, attachments, relations, inverse relations) previously inline
  in Worker.syncIssueDetailsBatch, including the pruneWhenComplete
  page-completeness composition and the completeness-set comments; returns
  clean = all five collections clean. Embedded-file extraction stays a
  best-effort sub-write inside the comment upsert closure via Deps.Extract.

- Extractor moves the embedded-file module (pure CDN-URL parse + HEAD/upsert
  I/O tail, formerly Worker.extractAndStoreEmbeddedFiles/fetchFileSize);
  nil HTTPClient defaults to http.DefaultClient, injectable for tests (the
  embeddedFileCache precedent). The moved store test now routes HEADs
  through an injected client instead of hitting the real CDN.

Behavior-preserving for the worker: the clean values are ignored for now,
and the touch/dequeue/gate logic in syncIssueDetailsBatch is unchanged.
CONTEXT.md's "Sync reconcile tail" entry updated for the new home and the
clean return.
